mm-iface-modem: mark signal strength as recent on init
When initializing the Modem interface, we set the signal strength
to the default value (0%) and we are certain this is the recent value
until the modem reports the actual signal strength if available.
Therefore, mark the signal strength as recent instead of cached.

(cherry picked from commit 3239fc22ff66819540830191f7b5462b0241713c)