cinterion: add retry mechanism to the ^SWWAN? command
Since it is possible that the delayed status request for the connection
(^SWWAN?) also does not return the updated status a retry is needed.
This retry is only done for the activation path. So the connection
status monitoring and the deactivation path is not affected.

Signed-off-by: Heiko Thiery <heiko.thiery@gmail.com>
(cherry picked from commit b5c9bae104e9397cb24e493954ef4c41ce0a89a3)