broadband-modem-qmi: use DSD system status to report packet service state
We now support modem-reported packet service state, instead of
guessing it in the 3GPP interface logic.

The DSD system status reporting a valid access technology is an
indication that the packet service is ready. Instead of holding back
the registration status until the DSD system status reports a valid
value, we will now report the modem as registered, and bind the DSD
system status exclusively to the packet service status.

In the simple connection attempt we're already waiting first for being
registered, and then for being attached in PS.