quectel: avoid incorrect module version report after the upgrade
After LVFS upgrade, the module needs to be restarted. ModemManager
queries the module information, sometimes the module is not ready, and
the information queried by "mmcli -m any --firmware-status" is
incomplete,

It is now modified to re query once per second when obtaining
incomplete versions, up to a maximum of 16 times.

Signed-off-by: mank.wang <mank.wang@quectel.com>