iface-modem: also run disable on failed low power mode update
If the user requested to set the modem in low power mode while
enabled, the user is also expecting the device to be implicitly
disabled.

If for any reason setting the modem to low power mode fails (e.g. it
could be not supported by the device), we should still try to run the
modem disabling sequence, as that is the expectation.