broadband-modem-mbim: add qdu firmware update check
The manufacturers can offer a firmware update file 'ota.bin' with which
the firmware can be updated in the modem via a manufacturer tool using
windows. The tool is only available for Windows and can therefore not be used
for embedded devices.

A investigation has shown that this tool updates the firmware via 'mbim-qdu'.
The update service 'fwupd' which has a 'modemmanager' plugin can install the
firmware via 'mbim-qdu'. For this to work, the ModemManager must export this
update method via the firmware interface.

As the 'mbim-qdu' is a generic update method, the check is integrated into the
'mm-broadband-modem-mbim' class. If 'mbim-qdu' update method is supported by
the mbim interface and the udev env variable 'ID_MM_MBIM_QDU' is set for this
modem, then the 'mbim-qdu' update is exported via dbus.

The following line must be added to the '77-mm-cinterion-port-types.rules'
udev rules file, for example, so that the mbim-qdu update for the 'MV31' can
be used.

ATTRS{vendor}=="0x1e2d", ATTRS{device}=="0x00b3", ENV{ID_MM_MBIM_QDU}="1"

This is the corresponding output of the 'mmcli' command to show the supported
update method for the 'MV31'.

root@G3-10940 ~ # mmcli -m Modem1 --firmware-status
  Firmware |     method: mbim-qdu
           | device ids: USB\VID_1E2D&PID_00B3&REV_0414&CARRIER_GCF
           |             USB\VID_1E2D&PID_00B3&REV_0414
           |             USB\VID_1E2D&PID_00B3
           |             USB\VID_1E2D
           |    version: T99W175.F0.1.0.0.9.GC.004.063 - 0A000804

Signed-off-by: Florian Eckert <fe@dev.tdt.de>